Refactor spacer block to share edit method with the web.
