+++ packages/block-library/src/spacer/edit.js
@@ -6,24 +6,28 @@ import classnames from 'classnames';
 /**
  * WordPress dependencies
  */
-import { useState } from '@wordpress/element';
 import { __ } from '@wordpress/i18n';
 import { InspectorControls } from '@wordpress/block-editor';
-import { BaseControl, PanelBody, ResizableBox } from '@wordpress/components';
+import { PanelBody, ResizableBox, RangeControl } from '@wordpress/components';
 import { compose, withInstanceId } from '@wordpress/compose';
 import { withDispatch } from '@wordpress/data';
+
+const minSpacerHeight = 20;
+const maxSpacerHeight = 500;
 
 const SpacerEdit = ( {
 	attributes,
 	isSelected,
 	setAttributes,
-	instanceId,
 	onResizeStart,
 	onResizeStop,
 } ) => {
 	const { height } = attributes;
-	const id = `block-spacer-height-input-${ instanceId }`;
-	const [ inputHeightValue, setInputHeightValue ] = useState( height );
+	const changeAttribute = ( value ) => {
+		setAttributes( {
+			height: value,
+		} );
+	};
 
 	return (
 		<>
@@ -37,7 +41,7 @@ const SpacerEdit = ( {
 				size={ {
 					height,
 				} }
-				minHeight="20"
+				minHeight={ minSpacerHeight }
 				enable={ {
 					top: false,
 					right: false,
@@ -48,6 +52,7 @@ const SpacerEdit = ( {
 					bottomLeft: false,
 					topLeft: false,
 				} }
+				isSelected={ isSelected }
 				onResizeStart={ onResizeStart }
 				onResizeStop={ ( event, direction, elt, delta ) => {
 					onResizeStop();
@@ -55,38 +60,19 @@ const SpacerEdit = ( {
 					setAttributes( {
 						height: spacerHeight,
 					} );
-					setInputHeightValue( spacerHeight );
 				} }
 			/>
 			<InspectorControls>
 				<PanelBody title={ __( 'Spacer settings' ) }>
-					<BaseControl label={ __( 'Height in pixels' ) } id={ id }>
-						<input
-							type="number"
-							id={ id }
-							onChange={ ( event ) => {
-								let spacerHeight = parseInt(
-									event.target.value,
-									10
-								);
-								setInputHeightValue( spacerHeight );
-								if ( isNaN( spacerHeight ) ) {
-									// Set spacer height to default size and input box to empty string
-									setInputHeightValue( '' );
-									spacerHeight = 100;
-								} else if ( spacerHeight < 20 ) {
-									// Set spacer height to minimum size
-									spacerHeight = 20;
-								}
-								setAttributes( {
-									height: spacerHeight,
-								} );
-							} }
-							value={ inputHeightValue }
-							min="20"
-							step="10"
-						/>
-					</BaseControl>
+					<RangeControl
+						label={ __( 'Height in pixels' ) }
+						min={ minSpacerHeight }
+						max={ maxSpacerHeight }
+						separatorType={ 'none' }
+						value={ height }
+						onChange={ changeAttribute }
+						step={ 10 }
+					/>
 				</PanelBody>
 			</InspectorControls>
 		</>
